+++ backend/core/src/listings/dto/listing.dto.ts
@@ -831,6 +831,16 @@ export class ListingFilterParams extends BaseFilter {
 
   @Expose()
   @ApiProperty({
+    type: Boolean,
+    example: "true",
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsBooleanString({ groups: [ValidationsGroupsEnum.default] })
+  [ListingFilterKeys.independentLivingHousing]?: boolean;
+
+  @Expose()
+  @ApiProperty({
     type: Number,
     example: "300",
     required: false,
@@ -947,6 +957,7 @@ export const filterTypeToFieldMap: Record<keysWithMappedField, string> = {
   bedrooms: "summaryUnitType.num_bedrooms",
   zipcode: "buildingAddress.zipCode",
   seniorHousing: "reservedCommunityType.name",
+  independentLivingHousing: "reservedCommunityType.name",
   // This is the inverse of the explanation for maxRent below.
   minRent: "unitsSummary.monthly_rent_max",
   // The maxRent filter uses the monthly_rent_min field to avoid missing units