+++ backend/core/src/listings/types/listing-filter-keys-enum.ts
@@ -6,6 +6,7 @@ export enum ListingFilterKeys {
   zipcode = "zipcode",
   availability = "availability",
   seniorHousing = "seniorHousing",
+  independentLivingHousing = "independentLivingHousing",
   minRent = "minRent",
   maxRent = "maxRent",
   minAmiPercentage = "minAmiPercentage",