+++ backend/core/src/shared/filter/custom_filters.ts
@@ -25,6 +25,26 @@ export function addSeniorHousingQuery(qb: WhereExpression, filterValue: string) 
   }
 }
 
+export function addIndependentLivingHousingQuery(qb: WhereExpression, filterValue: string) {
+  const whereParameterName = ListingFilterKeys.independentLivingHousing
+  const independentLivingCommunityType = "specialNeeds"
+  const reservedCommunityTypeColumnName = `LOWER(CAST(${
+    filterTypeToFieldMap[ListingFilterKeys.independentLivingHousing]
+  } as text))`
+  if (filterValue == "true") {
+    qb.andWhere(`(${reservedCommunityTypeColumnName} = LOWER(:${whereParameterName}))`, {
+      [whereParameterName]: independentLivingCommunityType,
+    })
+  } else if (filterValue == "false") {
+    qb.andWhere(
+      `(${reservedCommunityTypeColumnName} IS NULL OR ${reservedCommunityTypeColumnName} <> LOWER(:${whereParameterName}))`,
+      {
+        [whereParameterName]: independentLivingCommunityType,
+      }
+    )
+  }
+}
+
 export function addAvailabilityQuery(
   qb: WhereExpression,
   filterValue: AvailabilityFilterEnum,