+++ backend/core/src/shared/filter/index.ts
@@ -9,6 +9,7 @@ import {
   addSeniorHousingQuery,
   addAvailabilityQuery,
   addMinAmiPercentageFilter,
+  addIndependentLivingHousingQuery,
 } from "./custom_filters"
 
 /**
@@ -51,6 +52,9 @@ export function addFilters<FilterParams extends Array<any>, FilterFieldMap>(
       switch (filterKey) {
         case ListingFilterKeys.seniorHousing:
           addSeniorHousingQuery(qb, filterValue)
+          continue
+        case ListingFilterKeys.independentLivingHousing:
+          addIndependentLivingHousingQuery(qb, filterValue)
           continue
         case ListingFilterKeys.availability:
           addAvailabilityQuery(qb, filterValue as AvailabilityFilterEnum, includeNulls)